(#000) Initialization of the LMS Project
I created a complete scheme for the monorepo project, in which I use Angular as the frontend
framework with TypeScript (JavaScript), and in the backend, we use Java with the Spring
environment, including Spring Security, Spring Web, Spring Data JPA, and Lombok.  With
respect to service usage, we created a user and the respective secret keys for the project
in GCP. To deploy the backend, we use Railway, and consequently, the creation of a Dockerfile.
For the frontend deployment, we use Vercel.

- Created a monolithic architecture in the backend. On the other hand,
  we created a classic component architecture in the frontend. The
  structured subdirectories will change as we create the different
  models and representations of the entity diagrams for the use case
  definitions.
- Used the PNPM environment to create and aggregate libraries and
  different dependencies in the frontend.  In the backend, we use Maven
  with different Spring modules.
- Defined the properties needed to deploy and use the database in
  Supabase and the security environment. We intend to enhance this by
  implementing Swagger dependencies, for example, to restrict the
  endpoints and avoid 303 errors when running the project.
- In every part of the project, we implemented a Markdown file tree
  structure and created different environments to easily deploy to
  production.

The reason for creating this project is to enhance and provide an
open-source EMR environment to create a better experience in the
real-world operations of different hospitals and healthcare spaces. We
intend to improve upon the project created by pateljannat, which is created
in Python and executed in a different environment, but lacks efficiency, as
observed at https://github.com/frappe/lms.

Signed-off-by: Joseph Gallegos <jaremgallegosp@gmail.com>