(#001) Creation of the Documentation Template
In the 'docs' directory, we created a weebpage using Docusaurus and a Rust
-based 'bundler' script. This bundler verifies user identity and manages
various systems creations. Its purpose is to create a Typst document and
validate it against Typst Universe standards. We also addressed
modifications in pnpm files when creating the Docusaurus project within
the webpage part of 'docs', and made changes to README.md due to issues
with the assets banner file.

- A Rust bundler was created to verify user creation, define user
  information, and validate credentials.
- Resolved modifications for various packages within the EMR project,
  whose name was changed to 'skilling', because the current project is
  now a LMS. We also plan to create more extensive and precise
  documentation, defining objects and providing additional documentation
  for the API and models used in the project.
- Regarding the TODOs, we need to modify certain aspects, such as the
  Youtube GCP API models used in the project.
- We plan to integrate an asset file tree definition, as we've included
  various documents like diagrams, relationship models, and other
  elements that enhance the project and define the different parts we
  created and primarily verified.

Finally, we established the complete stack used in this monorepo and
implemented variuos aspects across different parts, including the use of
Jira and defining the Kanban board for this project. We aim to further
enhance package definitions and integrate the remaining TODO issues into
the project.

Signed-off-by: Joseph Gallegos <jaremgallegosp@gmail.com>