(#003) Enhance Typst templates and integrate DaisyUI
This commit focuses on improving the structure of our Typst documents and
creating a more comprenhensive template documentation. We've integrated
the DaisyUI dependency and removed MaterialUI to prevent library
conflicts. A significant change also includes defining the usage of SVG
icons, with their respective links now documented in the TODO file.

- Created complete documentation templates for the repository,
  integrating various Typst components and libraries.
- Switched the component library from MaterialUI to DaisyUI to enhance
  the project's design and functionality
- Updated the TODO file to include links to the documentation libraries
  used and tofinalize remaining tasks.

This commit primarly establishes the main project repository template on
Github and sets up the branches for the defined teams.

Signed-off-by: Joseh Gallegos <jaremgallegosp@gmail.com>